Moving CA1003 to its new location - no functional change. Note that this analyzer also covers CA1009 now and so I've add that rule to the props file as well.
