+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/CSharpUseGenericEventHandlerInstances.cs
@@ -1,19 +1,48 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System.Diagnostics;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 
 namespace Microsoft.ApiDesignGuidelines.Analyzers
-{                          
-    /// <summary>
-    /// CA1003: Use generic event handler instances
-    /// </summary>
+{
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
-    public sealed class CSharpUseGenericEventHandlerInstancesAnalyzer : UseGenericEventHandlerInstancesAnalyzer
+    public class CSharpUseGenericEventHandlerInstancesAnalyzer : UseGenericEventHandlerInstancesAnalyzer
     {
+        protected override AnalyzerBase GetAnalyzer(
+            Compilation compilation,
+            INamedTypeSymbol eventHandler,
+            INamedTypeSymbol genericEventHandler,
+            INamedTypeSymbol eventArgs,
+            INamedTypeSymbol comSourceInterfacesAttribute)
+        {
+            return new Analyzer(compilation, eventHandler, genericEventHandler, eventArgs, comSourceInterfacesAttribute);
+        }
 
+        private sealed class Analyzer : AnalyzerBase
+        {
+            public Analyzer(
+                Compilation compilation,
+                INamedTypeSymbol eventHandler,
+                INamedTypeSymbol genericEventHandler,
+                INamedTypeSymbol eventArgs,
+                INamedTypeSymbol comSourceInterfacesAttribute)
+                : base(compilation, eventHandler, genericEventHandler, eventArgs, comSourceInterfacesAttribute)
+            {
+            }
+
+            protected override bool IsViolatingEventHandler(INamedTypeSymbol type)
+            {
+                return !IsValidLibraryEventHandlerInstance(type);
+            }
+
+            protected override bool IsAssignableTo(Compilation compilation, ITypeSymbol fromSymbol, ITypeSymbol toSymbol)
+            {
+                return
+                    fromSymbol != null &&
+                    toSymbol != null &&
+                    ((CSharpCompilation)compilation).ClassifyConversion(fromSymbol, toSymbol).IsImplicit;
+            }
+        }
     }
 }
