+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.props
@@ -10,6 +10,8 @@
       <CodeAnalysisRuleSetOverrides>
         $(CodeAnalysisRuleSetOverrides);
         -Microsoft.Design#CA1001;
+        -Microsoft.Design#CA1003;
+        -Microsoft.Design#CA1009;
         -Microsoft.Design#CA1019;
 
         -Microsoft.Globalization#CA1309;
