+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/VisualBasic/Microsoft.ApiDesignGuidelines.VisualBasic.Analyzers.props
@@ -10,8 +10,10 @@
       <CodeAnalysisRuleSetOverrides>
         $(CodeAnalysisRuleSetOverrides);
         -Microsoft.Design#CA1001;
+        -Microsoft.Design#CA1003;
+        -Microsoft.Design#CA1009;
         -Microsoft.Design#CA1019;
-        
+
         -Microsoft.Globalization#CA1309;
       </CodeAnalysisRuleSetOverrides>
    </PropertyGroup>
